+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## dev
 
+## Beta (0.0.11)
+
+- Fix error in `DependencyParserCoreNLP` that returning a constituency tree rather than a dependency tree.
+- `ConstituencyParserCoreNLP` now use CoreNLP like `DependencyParserCoreNLP` instead of Stanza.
+
 ## Beta (0.0.10)
 
 - Fix import problem with CACHE.